+++ wallets/cosmos-extension-metamask/src/cosmos-metamask-extension.ts
@@ -1,0 +1,5 @@
+import { cosmosSnapExtensionInfo, CosmosMetamaskExtensionWallet } from './extension';
+
+const cosmosExtensionMetamaskExtension = new CosmosMetamaskExtensionWallet(cosmosSnapExtensionInfo);
+
+export const wallets = [cosmosExtensionMetamaskExtension];