+++ wallets/cosmos-extension-metamask/src/extension/types.ts
@@ -1,0 +1,8 @@
+export interface Key {
+  readonly name: string;
+  readonly algo: string;
+  readonly pubKey: Uint8Array;
+  readonly address: Uint8Array;
+  readonly bech32Address: string;
+  readonly isNanoLedger: boolean;
+}