+++ wallets/leap-metamask-cosmos-snap/src/extension/registry.ts
@@ -1,13 +1,13 @@
 import { Wallet } from '@cosmos-kit/core';
 
-import { ICON } from '../constant';
+import { LEAP_LOGO, METAMASK_LOGO } from '../constant';
 
 export const metamaskCosmosSnapInfo: Wallet = {
   name: 'leap-metamask-cosmos-snap',
   description:
     'The Leap Cosmos MetaMask Snap allows signing of Cosmos transactions using MetaMask.',
-  prettyName: 'Leap Cosmos MetaMask Snap',
-  logo: ICON,
+  prettyName: 'Leap Cosmos MetaMask',
+  logo: { major: METAMASK_LOGO, minor: LEAP_LOGO },
   mode: 'extension',
   extends: 'MetaMask',
   mobileDisabled: () => !('ethereum' in window),