+++ packages/core/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.3](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/core@2.6.2...@cosmos-kit/core@2.6.3) (2023-09-15)
+
+**Note:** Version bump only for package @cosmos-kit/core
+
 ## [2.6.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/core@2.6.1...@cosmos-kit/core@2.6.2) (2023-09-12)
 
 **Note:** Version bump only for package @cosmos-kit/core