+++ packages/docs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.7](https://github.com/cosmology-tech/cosmos-kit/compare/docs@2.5.6...docs@2.5.7) (2023-09-15)
+
+**Note:** Version bump only for package docs
+
+## [2.5.6](https://github.com/cosmology-tech/cosmos-kit/compare/docs@2.5.5...docs@2.5.6) (2023-09-13)
+
+**Note:** Version bump only for package docs
+
+## [2.5.5](https://github.com/cosmology-tech/cosmos-kit/compare/docs@2.5.4...docs@2.5.5) (2023-09-13)
+
+**Note:** Version bump only for package docs
+
+## [2.5.4](https://github.com/cosmology-tech/cosmos-kit/compare/docs@2.5.3...docs@2.5.4) (2023-09-13)
+
+**Note:** Version bump only for package docs
+
 ## [2.5.3](https://github.com/cosmology-tech/cosmos-kit/compare/docs@2.5.2...docs@2.5.3) (2023-09-12)
 
 **Note:** Version bump only for package docs